requirements.txt
gTTS
moviepy==1.0.3
Pillow==9.5.0
numpy
requests
pandas
openpyxl